+++ src/main/kotlin/com/nylas/models/FindEventQueryParams.kt
@@ -11,6 +11,13 @@ data class FindEventQueryParams(
    */
   @Json(name = "calendar_id")
   val calendarId: String,
+  /**
+   * Specify fields that you want Nylas to return, as a comma-separated list (for example, select=id,updated_at).
+   * This allows you to receive only the portion of object data that you're interested in.
+   * You can use select to optimize response size and reduce latency by limiting queries to only the information that you need
+   */
+  @Json(name = "select")
+  var select: String? = null,
 ) : IQueryParams {
   /**
    * Builder for [FindEventQueryParams].
@@ -19,12 +26,22 @@ data class FindEventQueryParams(
   data class Builder(
     private val calendarId: String,
   ) {
+    private var select: String? = null
+
+    /**
+     * Sets the fields to select.
+     * @param select The fields to select.
+     * @return The builder.
+     */
+    fun select(select: String?) = apply { this.select = select }
+
     /**
      * Builds a new [FindEventQueryParams] instance.
      * @return [FindEventQueryParams]
      */
     fun build() = FindEventQueryParams(
       calendarId,
+      select,
     )
   }
 }