+++ src/main/kotlin/com/nylas/models/ListCalendersQueryParams.kt
@@ -23,6 +23,13 @@ data class ListCalendersQueryParams(
    */
   @Json(name = "metadata_pair")
   val metadataPair: Map<String, String>? = null,
+  /**
+   * Specify fields that you want Nylas to return, as a comma-separated list (for example, select=id,updated_at).
+   * This allows you to receive only the portion of object data that you're interested in.
+   * You can use select to optimize response size and reduce latency by limiting queries to only the information that you need
+   */
+  @Json(name = "select")
+  var select: String? = null,
 ) : IQueryParams {
   /**
    * Builder for [ListCalendersQueryParams].
@@ -31,6 +38,7 @@ data class ListCalendersQueryParams(
     private var limit: Int? = null
     private var pageToken: String? = null
     private var metadataPair: Map<String, String>? = null
+    private var select: String? = null
 
     /**
      * Sets the maximum number of objects to return.
@@ -56,6 +64,13 @@ data class ListCalendersQueryParams(
     fun metadataPair(metadataPair: Map<String, String>?) = apply { this.metadataPair = metadataPair }
 
     /**
+     * Set the fields to return in the response.
+     * @param select List of field names to return (e.g. "id,grant_id,name")
+     * @return The builder.
+     */
+    fun select(select: String?) = apply { this.select = select }
+
+    /**
      * Builds a [ListCalendersQueryParams] instance.
      * @return The [ListCalendersQueryParams] instance.
      */
@@ -63,6 +78,7 @@ data class ListCalendersQueryParams(
       limit,
       pageToken,
       metadataPair,
+      select,
     )
   }
 }