+++ src/main/kotlin/com/nylas/models/ListEventQueryParams.kt
@@ -124,6 +124,13 @@ data class ListEventQueryParams(
    */
   @Json(name = "event_type")
   val eventType: EventType? = null,
+  /**
+   * Specify fields that you want Nylas to return, as a comma-separated list (for example, select=id,updated_at).
+   * This allows you to receive only the portion of object data that you're interested in.
+   * You can use select to optimize response size and reduce latency by limiting queries to only the information that you need
+   */
+  @Json(name = "select")
+  var select: String? = null,
 ) : IQueryParams {
   /**
    * Builder for [ListEventQueryParams].
@@ -152,6 +159,7 @@ data class ListEventQueryParams(
     private var updatedAfter: Long? = null
     private var attendees: List<String>? = null
     private var eventType: EventType? = null
+    private var select: String? = null
 
     /**
      * Sets the maximum number of objects to return.
@@ -300,6 +308,13 @@ data class ListEventQueryParams(
     fun eventType(eventType: EventType?) = apply { this.eventType = eventType }
 
     /**
+     * Set the fields to return in the response.
+     * @param select List of field names to return (e.g. "id,grant_id,title")
+     * @return The builder.
+     */
+    fun select(select: String?) = apply { this.select = select }
+
+    /**
      * Builds a [ListEventQueryParams] instance.
      * @return The [ListEventQueryParams] instance.
      */
@@ -323,6 +338,7 @@ data class ListEventQueryParams(
       updatedAfter,
       attendees,
       eventType,
+      select,
     )
   }
 }